fix: Table not sorting on mutation successful.
