+++ src/hooks/useTableSort.ts
@@ -1,22 +1,16 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import { useEffect, useState } from 'react'
 import { getCorrectType } from '@/utils/utils'
-import { AmbigiousObject, LendingTableSort } from '../pages/Lending/types/LendingTypes'
+import { AmbigiousObject, LendingTableSort } from '@/components/ui/table'
 
 const useLendingTableSort = <T extends AmbigiousObject>({ data }: { data: T[] }) => {
   const [sortBy, setSortBy] = useState<LendingTableSort<T>>({ columnName: undefined, orderBy: 'none' })
-  const [originalData, setOriginalData] = useState<T[]>(data)
   const [sortedData, setSortedData] = useState<T[]>(data)
 
   useEffect(() => {
-    if (!originalData.length) {
-      setOriginalData(data)
-    }
-
-    if (!sortedData.length) {
-      setSortedData(data)
-    }
+    handleSort()
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data])
+  }, [sortBy.columnName, sortBy.orderBy, data])
 
   function dataSetter(data: T[]) {
     if (sortBy.orderBy === 'desc') {
@@ -29,36 +23,31 @@ const useLendingTableSort = <T extends AmbigiousObject>({ data }: { data: T[] })
   function handleSort() {
     const { orderBy, columnName } = sortBy
     if (orderBy === 'none' || !columnName) {
-      setSortedData(originalData)
+      setSortedData(data)
       return
     }
     const columnType = typeof getCorrectType(sortedData[0]?.[columnName])
     if (!columnType) {
-      setSortedData(originalData)
+      setSortedData(data)
       return
     }
     let _result
     if (columnType === 'number') {
-      _result = [...originalData].sort((a, b) => {
+      _result = [...data].sort((a, b) => {
         return ((a[columnName] as number) - b[columnName]) as number
       })
     } else if (columnType === 'string') {
-      _result = [...originalData].sort((a, b) => {
+      _result = [...data].sort((a, b) => {
         return (a[columnName] as string).localeCompare(b[columnName] as string)
       })
     } else if (columnType === 'object') {
-      _result = [...originalData].sort((a, b) => {
+      _result = [...data].sort((a, b) => {
         return new Date(a[columnName] as string).getTime() - new Date(b[columnName] as string).getTime()
       })
     }
 
-    dataSetter(_result || originalData)
+    dataSetter(_result || data)
   }
-
-  useEffect(() => {
-    handleSort()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [sortBy.columnName, sortBy.orderBy])
 
   return {
     handleSort,
@@ -66,7 +55,6 @@ const useLendingTableSort = <T extends AmbigiousObject>({ data }: { data: T[] })
     setSortedData,
     sortBy,
     setSortBy,
-    setOriginalData,
   }
 }
